# contracts/openapi/keynetra-v0.1.2.yaml
openapi: 3.0.3
info:
  title: KeyNetra
  version: 0.1.2
paths:
...
  /policies:
    get:
      tags:
      - management
      summary: List Policies
      operationId: list_policies
      security:
      - HTTPBearer: []
      - APIKeyHeader: []
      parameters:
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          default: 50
          title: Limit
      - name: cursor
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - nullable: true
          title: Cursor
      - $ref: '#/components/parameters/ApiVersionHeader'
      - $ref: '#/components/parameters/TenantHeader'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse_list_PolicyOut__'
              examples:
                success:
                  summary: Standard success envelope
                  value:
                    data: {}
                    meta:
                      request_id: req_example_123
                      limit: null
                      next_cursor: null
                      extra: {}
                    error: null
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestIdHeader'
            X-API-Version:
              $ref: '#/components/headers/ApiVersionHeader'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                error:
                  summary: Standard error envelope
                  value:
                    data: null
                    meta:
                      request_id: req_example_123
                      limit: null
                      next_cursor: null
                      extra: {}
                    error:
                      code: validation_error
                      message: request failed
                      details: null
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestIdHeader'
            X-API-Version:
              $ref: '#/components/headers/ApiVersionHeader'
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '409':
          $ref: '#/components/responses/Error409'
        '429':
          $ref: '#/components/responses/Error429'
        '500':
          $ref: '#/components/responses/Error500'
    post:
      tags:
      - management
      summary: Create Policy
      operationId: create_policy
      security:
      - HTTPBearer: []
      - APIKeyHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PolicyCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse_PolicyOut_'
              examples:
                success:
                  summary: Standard success envelope
                  value:
                    data: {}
                    meta:
                      request_id: req_example_123
                      limit: null
                      next_cursor: null
                      extra: {}
                    error: null
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestIdHeader'
            X-API-Version:
              $ref: '#/components/headers/ApiVersionHeader'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                error:
                  summary: Standard error envelope
                  value:
                    data: null
                    meta:
                      request_id: req_example_123
                      limit: null
                      next_cursor: null
                      extra: {}
                    error:
                      code: validation_error
                      message: request failed
                      details: null
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestIdHeader'
            X-API-Version:
              $ref: '#/components/headers/ApiVersionHeader'
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '409':
          $ref: '#/components/responses/Error409'
        '429':
          $ref: '#/components/responses/Error429'
        '500':
          $ref: '#/components/responses/Error500'
      parameters:
      - $ref: '#/components/parameters/ApiVersionHeader'
      - $ref: '#/components/parameters/TenantHeader'
  /policies/{policy_key}:
    put:
      tags:
      - management
      summary: Update Policy
      operationId: update_policy
      security:
      - HTTPBearer: []
      - APIKeyHeader: []
      parameters:
      - name: policy_key
        in: path
        required: true
        schema:
          type: string
          title: Policy Key
      - $ref: '#/components/parameters/ApiVersionHeader'
      - $ref: '#/components/parameters/TenantHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PolicyCreate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse_PolicyOut_'
              examples:
                success:
                  summary: Standard success envelope
                  value:
                    data: {}
                    meta:
                      request_id: req_example_123
                      limit: null
                      next_cursor: null
                      extra: {}
                    error: null
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestIdHeader'
            X-API-Version:
              $ref: '#/components/headers/ApiVersionHeader'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                error:
                  summary: Standard error envelope
                  value:
                    data: null
                    meta:
                      request_id: req_example_123
                      limit: null
                      next_cursor: null
                      extra: {}
                    error:
                      code: validation_error
                      message: request failed
                      details: null
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestIdHeader'
            X-API-Version:
              $ref: '#/components/headers/ApiVersionHeader'
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '409':
          $ref: '#/components/responses/Error409'
        '429':
          $ref: '#/components/responses/Error429'
        '500':
          $ref: '#/components/responses/Error500'
    delete:
      tags:
      - management
      summary: Delete Policy
      operationId: delete_policy
      security:
      - HTTPBearer: []
      - APIKeyHeader: []
      parameters:
      - name: policy_key
        in: path
        required: true
        schema:
          type: string
          title: Policy Key
      - $ref: '#/components/parameters/ApiVersionHeader'
      - $ref: '#/components/parameters/TenantHeader'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse_dict_str__str__'
              examples:
                success:
                  summary: Standard success envelope
                  value:
                    data: {}
                    meta:
                      request_id: req_example_123
                      limit: null
                      next_cursor: null
                      extra: {}
                    error: null
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestIdHeader'
            X-API-Version:
              $ref: '#/components/headers/ApiVersionHeader'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                error:
                  summary: Standard error envelope
                  value:
                    data: null
                    meta:
                      request_id: req_example_123
                      limit: null
                      next_cursor: null
                      extra: {}
                    error:
                      code: validation_error
                      message: request failed
                      details: null
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestIdHeader'
            X-API-Version:
              $ref: '#/components/headers/ApiVersionHeader'
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '409':
          $ref: '#/components/responses/Error409'
        '429':
          $ref: '#/components/responses/Error429'
        '500':
          $ref: '#/components/responses/Error500'
  /policies/dsl:
    post:
      tags:
      - management
      summary: Create Policy From Dsl
      operationId: create_policy_from_dsl
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PolicyDslCreate'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse_PolicyOut_'
              examples:
                success:
                  summary: Standard success envelope
                  value:
                    data: {}
                    meta:
                      request_id: req_example_123
                      limit: null
                      next_cursor: null
                      extra: {}
                    error: null
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestIdHeader'
            X-API-Version:
              $ref: '#/components/headers/ApiVersionHeader'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                error:
                  summary: Standard error envelope
                  value:
                    data: null
                    meta:
                      request_id: req_example_123
                      limit: null
                      next_cursor: null
                      extra: {}
                    error:
                      code: validation_error
                      message: request failed
                      details: null
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestIdHeader'
            X-API-Version:
              $ref: '#/components/headers/ApiVersionHeader'
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '409':
          $ref: '#/components/responses/Error409'
        '429':
          $ref: '#/components/responses/Error429'
        '500':
          $ref: '#/components/responses/Error500'
      security:
      - HTTPBearer: []
      - APIKeyHeader: []
      parameters:
      - $ref: '#/components/parameters/ApiVersionHeader'
      - $ref: '#/components/parameters/TenantHeader'
  /policies/{policy_key}/rollback/{version}:
    post:
      tags:
      - management
      summary: Rollback Policy
      operationId: rollback_policy
      security:
      - HTTPBearer: []
      - APIKeyHeader: []
      parameters:
      - name: policy_key
        in: path
        required: true
        schema:
          type: string
          title: Policy Key
      - name: version
        in: path
        required: true
        schema:
          type: integer
          title: Version
      - $ref: '#/components/parameters/ApiVersionHeader'
      - $ref: '#/components/parameters/TenantHeader'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse_dict_str__Union_int__str___'
              examples:
                success:
                  summary: Standard success envelope
                  value:
                    data: {}
                    meta:
                      request_id: req_example_123
                      limit: null
                      next_cursor: null
                      extra: {}
                    error: null
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestIdHeader'
            X-API-Version:
              $ref: '#/components/headers/ApiVersionHeader'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                error:
                  summary: Standard error envelope
                  value:
                    data: null
                    meta:
                      request_id: req_example_123
                      limit: null
                      next_cursor: null
                      extra: {}
                    error:
                      code: validation_error
                      message: request failed
                      details: null
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestIdHeader'
            X-API-Version:
              $ref: '#/components/headers/ApiVersionHeader'
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '409':
          $ref: '#/components/responses/Error409'
        '429':
          $ref: '#/components/responses/Error429'
        '500':
          $ref: '#/components/responses/Error500'
...
components:
  schemas:
...
    MetaBody:
      properties:
        request_id:
          anyOf:
          - type: string
          - nullable: true
          title: Request Id
        limit:
          anyOf:
          - type: integer
          - nullable: true
          title: Limit
        next_cursor:
          anyOf:
          - type: string
          - nullable: true
          title: Next Cursor
        extra:
          additionalProperties: true
          type: object
          title: Extra
      additionalProperties: false
      type: object
      title: MetaBody
...
    PolicyCreate:
      properties:
        action:
          type: string
          title: Action
        effect:
          type: string
          title: Effect
          default: allow
        priority:
          type: integer
          title: Priority
          default: 100
        state:
          type: string
          title: State
          default: active
        conditions:
          additionalProperties: true
          type: object
          title: Conditions
      additionalProperties: false
      type: object
      required:
      - action
      title: PolicyCreate
    PolicyDslCreate:
      properties:
        dsl:
          type: string
          title: Dsl
      additionalProperties: false
      type: object
      required:
      - dsl
      title: PolicyDslCreate
    PolicyOut:
      properties:
        id:
          type: integer
          title: Id
        action:
          type: string
          title: Action
        effect:
          type: string
          title: Effect
        priority:
          type: integer
          title: Priority
        state:
          type: string
          title: State
          default: active
        conditions:
          additionalProperties: true
          type: object
          title: Conditions
      additionalProperties: false
      type: object
      required:
      - id
      - action
      - effect
      - priority
      - conditions
      title: PolicyOut
...
    SuccessResponse_PolicyOut_:
      properties:
        data:
          $ref: '#/components/schemas/PolicyOut'
        meta:
          $ref: '#/components/schemas/MetaBody'
        error:
          allOf:
          - $ref: '#/components/schemas/ErrorBody'
          nullable: true
          title: Error
      additionalProperties: false
      type: object
      required:
      - data
      title: SuccessResponse[PolicyOut]
...
    SuccessResponse_dict_str__Union_int__str___:
      properties:
        data:
          additionalProperties:
            anyOf:
            - type: integer
            - type: string
          type: object
          title: Data
        meta:
          $ref: '#/components/schemas/MetaBody'
        error:
          allOf:
          - $ref: '#/components/schemas/ErrorBody'
          nullable: true
          title: Error
      additionalProperties: false
      type: object
      required:
      - data
      title: SuccessResponse[dict[str, Union[int, str]]]
...
    SuccessResponse_dict_str__str__:
      properties:
        data:
          additionalProperties:
            type: string
          type: object
          title: Data
        meta:
          $ref: '#/components/schemas/MetaBody'
        error:
          allOf:
          - $ref: '#/components/schemas/ErrorBody'
          nullable: true
          title: Error
      additionalProperties: false
      type: object
      required:
      - data
      title: SuccessResponse[dict[str, str]]
...
    SuccessResponse_list_PolicyOut__:
      properties:
        data:
          items:
            $ref: '#/components/schemas/PolicyOut'
          type: array
          title: Data
        meta:
          $ref: '#/components/schemas/MetaBody'
        error:
          allOf:
          - $ref: '#/components/schemas/ErrorBody'
          nullable: true
          title: Error
      additionalProperties: false
      type: object
      required:
      - data
      title: SuccessResponse[list[PolicyOut]]
...
    ApiErrorCode:
      title: ApiErrorCode
      type: string
      enum:
      - bad_request
      - unauthorized
      - forbidden
      - not_found
      - conflict
      - too_many_requests
      - validation_error
      - database_error
      - internal_error
    ErrorBody:
      title: ErrorBody
      type: object
      required:
      - code
      - message
      properties:
        code:
          $ref: '#/components/schemas/ApiErrorCode'
        message:
          title: Message
          type: string
        details:
          title: Details
          nullable: true
    ErrorResponse:
      title: ErrorResponse
      type: object
      required:
      - error
      properties:
        data:
          nullable: true
          title: Data
        meta:
          $ref: '#/components/schemas/MetaBody'
        error:
          $ref: '#/components/schemas/ErrorBody'
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer
    APIKeyHeader:
      type: apiKey
      in: header
      name: X-API-Key
  parameters:
    ApiVersionHeader:
      name: X-API-Version
      in: header
      required: false
      schema:
        type: string
        default: v1
      description: 'Requested API version. Supported values: v1.'
    TenantHeader:
      name: X-Tenant-Id
      in: header
      required: false
      schema:
        type: string
      description: Tenant key for strict tenancy and management operations.
  headers:
    ApiVersionHeader:
      description: Resolved API version for the response.
      schema:
        type: string
        default: v1
    RequestIdHeader:
      description: Correlation id for the request.
      schema:
        type: string
  responses:
    Error400:
      description: Bad Request
      content:
        application/json:
          schema: &id001
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            error:
              summary: Standard error envelope
              value:
                data: null
                meta:
                  request_id: req_example_123
                  limit: null
                  next_cursor: null
                  extra: {}
                error:
                  code: bad_request
                  message: request failed
                  details: null
      headers:
        X-Request-Id:
          $ref: '#/components/headers/RequestIdHeader'
        X-API-Version:
          $ref: '#/components/headers/ApiVersionHeader'
    Error401:
      description: Unauthorized
      content:
        application/json:
          schema: *id001
          examples:
            error:
              summary: Standard error envelope
              value:
                data: null
                meta:
                  request_id: req_example_123
                  limit: null
                  next_cursor: null
                  extra: {}
                error:
                  code: unauthorized
                  message: request failed
                  details: null
      headers:
        X-Request-Id:
          $ref: '#/components/headers/RequestIdHeader'
        X-API-Version:
          $ref: '#/components/headers/ApiVersionHeader'
    Error403:
      description: Forbidden
      content:
        application/json:
          schema: *id001
          examples:
            error:
              summary: Standard error envelope
              value:
                data: null
                meta:
                  request_id: req_example_123
                  limit: null
                  next_cursor: null
                  extra: {}
                error:
                  code: forbidden
                  message: request failed
                  details: null
      headers:
        X-Request-Id:
          $ref: '#/components/headers/RequestIdHeader'
        X-API-Version:
          $ref: '#/components/headers/ApiVersionHeader'
    Error404:
      description: Not Found
      content:
        application/json:
          schema: *id001
          examples:
            error:
              summary: Standard error envelope
              value:
                data: null
                meta:
                  request_id: req_example_123
                  limit: null
                  next_cursor: null
                  extra: {}
                error:
                  code: not_found
                  message: request failed
                  details: null
      headers:
        X-Request-Id:
          $ref: '#/components/headers/RequestIdHeader'
        X-API-Version:
          $ref: '#/components/headers/ApiVersionHeader'
    Error409:
      description: Conflict
      content:
        application/json:
          schema: *id001
          examples:
            error:
              summary: Standard error envelope
              value:
                data: null
                meta:
                  request_id: req_example_123
                  limit: null
                  next_cursor: null
                  extra: {}
                error:
                  code: conflict
                  message: request failed
                  details: null
      headers:
        X-Request-Id:
          $ref: '#/components/headers/RequestIdHeader'
        X-API-Version:
          $ref: '#/components/headers/ApiVersionHeader'
...
    Error429:
      description: Too Many Requests
      content:
        application/json:
          schema: *id001
          examples:
            error:
              summary: Standard error envelope
              value:
                data: null
                meta:
                  request_id: req_example_123
                  limit: null
                  next_cursor: null
                  extra: {}
                error:
                  code: too_many_requests
                  message: request failed
                  details: null
      headers:
        X-Request-Id:
          $ref: '#/components/headers/RequestIdHeader'
        X-API-Version:
          $ref: '#/components/headers/ApiVersionHeader'
    Error500:
      description: Internal Server Error
      content:
        application/json:
          schema: *id001
          examples:
            error:
              summary: Standard error envelope
              value:
                data: null
                meta:
                  request_id: req_example_123
                  limit: null
                  next_cursor: null
                  extra: {}
                error:
                  code: internal_error
                  message: request failed
                  details: null
      headers:
        X-Request-Id:
          $ref: '#/components/headers/RequestIdHeader'
        X-API-Version:
          $ref: '#/components/headers/ApiVersionHeader'
servers:
- url: /
  description: Relative base URL. Configure the SDK/client host per environment.